moved dependencies to redbox-core
